It looks like a really big change, but it's mostly because the loader flags are now a uint64_t instead of an int.  We need more flag room.
By increasing the size of the loader flags, I was able to remove the noipv4
and noipv6 members from the loaderData_s struct and the networkDeviceConfig
struct.  I now just use the flags throughout.

This change will skip the Configure TCP/IP screen in loader if any of the
boot parameter pairs are passed:

   ip=<val> ipv6=<val>
   ip=<val> noipv6
   ipv6=<val> noipv4
   noipv4 noipv6

This brings back the similar feature we had in RHEL4 and pump-era FC releases
that let the loader skip over network configuration with ip=dhcp.

+++ loader2/driverdisk.c
@@ -47,7 +47,7 @@
 #include "../isys/eddsupport.h"
 
 /* boot flags */
-extern int flags;
+extern uint64_t flags;
 
 static char * driverDiskFiles[] = { "modinfo", "modules.dep", 
                                     "modules.cgz", "modules.alias", NULL };
